requirements.txt
plotly==5.11.0
streamlit==1.15.0
PyGithub == 1.55